- Added the temp file - Seperate upload module - Server.js implemented upload module
