+++ src/server/server.js
@@ -1,17 +1,12 @@
 (function () {
-    var app, fs, config, serveStatic, AWS, s3, express;
+    var app, fs, config, serveStatic, express, utils;
     express = require('express');
 
     app = express();
     fs = require('fs');
     serveStatic = require('serve-static');
     config = require('./config');
-    AWS = require('AWS-sdk');
 
-    // Create S3 Client object
-    var credentials = new AWS.SharedIniFileCredentials({ profile: config.awsProfile });
-    AWS.config.credentials = credentials;
-    s3 = new AWS.S3();
 
     app.use(serveStatic(config.bowerPath));
     app.use(serveStatic(config.rootPath + "/client"));
@@ -19,34 +14,17 @@
     // Configure the bucket Name
     app.get('/upload', function (req, res) {
         // Local Path
-        var filename = 'hello-world.txt';
-        var localPath = __dirname + '/../client/media/' + filename;
+        var localPath = __dirname + '/../client/media/hello-world.txt';
 
-        // S3 Key path
-        var keyName = '1/' + filename;
-
-        // Create fileStream to localPath
-        fs.readFile(localPath, function (err, data) {
-            if (err) {
-                console.log(err);
-                res.status(404).send(err);
-            }
-
-            var base64Data = new Buffer(data, 'binary');
-
-            var params = { Bucket: config.bucketName, Key: keyName, Body: base64Data, ACL: 'public-read' };
-            s3.putObject(params, function (err, data) {
-                if (err) {
-                    console.log(err)
-                    res.status(404).send(err);
-                }
-                else {
-                    console.log("Successfully uploaded data to http://s3.amazonaws.com/" + bucketName + "/" + keyName);
-                    res.status(200).send('success');
-                }
-            });
-        });
+		require('./utils')(localPath, function(err, result){
+			if(err) {
+				console.log(err);
+				res.status(404).send(err);
+			}
+			res.status(200).send(result);
+		});
     });
+
     app.listen(process.env.PORT || 3000);
     console.log("Listening on http://localhost:" + (process.env.PORT || '3000'));
